+++ resources/views/classrooms/index.blade.php
@@ -39,7 +39,7 @@
         @if($classroom->pivot->role == 2)
         <span class="admin">
           @yield('state')
-          <!-- <clone-classroom></clone-classroom> -->
+          <clone-classroom :classroom="{{ $classroom }}"></clone-classroom>
           <a href="/classroom/{{ $classroom->code }}/edit" class="has-text-white" {{ Popper::pop(__('menu.config')) }}><i class="far fa-cog"></i></a>
           <span {{ Popper::pop(__('classroom.admin')) }}>
             <i class="far fa-user-crown"></i>